+++ bin/publishCentral.sh
@@ -1,4 +1,5 @@
 #!/usr/bin/env bash
+set -e
 
 . $TRAVIS_BUILD_DIR/bin/setup.sh