+++ core/src/main/scala/org/http4s/headers/Set-Cookie.scala
@@ -1,10 +1,23 @@
 package org.http4s
 package headers
 
+import cats.data.NonEmptyList
+import scala.annotation.tailrec
 import org.http4s.parser.HttpHeaderParser
 import org.http4s.util.Writer
 
-object `Set-Cookie` extends HeaderKey.Internal[`Set-Cookie`] with HeaderKey.Singleton {
+object `Set-Cookie` extends HeaderKey.Internal[`Set-Cookie`] {
+  def from(headers: Headers): List[`Set-Cookie`] =
+    headers.toList.map(matchHeader).collect {
+      case Some(h) => h
+    }
+
+  def unapply(headers: Headers): Option[NonEmptyList[`Set-Cookie`]] =
+    from(headers) match {
+      case Nil => None
+      case h :: t => Some(NonEmptyList(h, t))
+    }
+
   override def parse(s: String): ParseResult[`Set-Cookie`] =
     HttpHeaderParser.SET_COOKIE(s)
 }