+++ core/src/main/scala/org/http4s/parser/Rfc2616BasicRules.scala
@@ -22,6 +22,7 @@ import cats._
 import cats.implicits._
 import org.http4s.{ParseFailure, ParseResult}
 import org.http4s.internal.parboiled2._
+import org.http4s.internal.parboiled2.CharPredicate.{ HexDigit => HEXDIG }
 
 // direct implementation of http://www.w3.org/Protocols/rfc2616/rfc2616-sec2.html#sec2
 private[http4s] trait Rfc2616BasicRules extends Parser {
@@ -76,12 +77,6 @@ private[http4s] trait Rfc2616BasicRules extends Parser {
   def ListSep = rule { oneOrMore("," ~ OptWS) }
 
   def LASTCHAR: Rule1[Char] = rule { push(input.charAt(cursor - 1)) }
-
-  // we don't match scoped IPv6 addresses
-  def IPv6Address = rule { oneOrMore(Hex | anyOf(":.")) }
-
-  def IPv6Reference: Rule1[String] = rule { capture("[" ~ IPv6Address ~ "]") }
-  // scalastyle:on public.methods.have.type
 }
 
 private[http4s] object Rfc2616BasicRules {