+++ server/src/main/scala/org/http4s/server/middleware/ResponseLogger.scala
@@ -6,6 +6,7 @@ import cats.effect._
 import cats.implicits._
 import fs2._
 import org.log4s.getLogger
+import scala.concurrent.ExecutionContext
 
 /**
   * Simple Middleware for Logging Responses As They Are Processed
@@ -13,13 +14,27 @@ import org.log4s.getLogger
 object ResponseLogger {
   private[this] val logger = getLogger
 
-  def apply[F[_]: Effect](logHeaders: Boolean, logBody: Boolean)
-                         (service: HttpService[F]): HttpService[F] = Service.lift { req =>
-    Stream.eval(service(req))
-      .map(_.toOption)
-      .unNone
-      .through(Logger.messageLogPipe[F, Response[F]](logHeaders, logBody)(logger))
-      .runLast
-      .map(_.getOrElse(Pass()))
-  }
+  def apply[F[_]](logHeaders: Boolean, logBody: Boolean)
+           (service: HttpService[F])
+           (implicit ec: ExecutionContext = ExecutionContext.global): HttpService[F] =
+    Service.lift { req => service(req).flatMap {
+        case Pass => Pass.now
+        case response: Response =>
+          if (!logBody) {
+            Logger.logMessage(response)(logHeaders, logBody)(logger)(strategy) >> Task(response)
+          } else {
+            fs2.async.unboundedQueue[Task, Byte].map { queue =>
+              val newBody = Stream.eval(queue.size.get)
+                .flatMap(size => queue.dequeue.take(size.toLong))
+
+              response.copy(
+                body = response.body
+                  .observe(queue.enqueue)
+                  .onFinalize {
+                    Logger.logMessage(response.copy(body = newBody))(logHeaders, logBody)(logger)(strategy)
+                  }
+              )
+            }
+          }
+    }}
 }