+++ servlet/src/main/scala/org/http4s/servlet/Http4sServlet.scala
@@ -10,7 +10,6 @@ import cats.implicits._
 import fs2.async
 import org.http4s.headers.`Transfer-Encoding`
 import org.http4s.server._
-import org.http4s.util.threads.DefaultExecutionContext
 import org.log4s.getLogger
 
 import scala.collection.JavaConverters._
@@ -19,7 +18,7 @@ import scala.concurrent.duration.Duration
 
 class Http4sServlet[F[_]](service: HttpService[F],
                           asyncTimeout: Duration = Duration.Inf,
-                          implicit private[this] val executionContext: ExecutionContext = DefaultExecutionContext,
+                          implicit private[this] val executionContext: ExecutionContext = ExecutionContext.global,
                           private[this] var servletIo: ServletIo[F])
                          (implicit F: Effect[F]) extends HttpServlet {
   private[this] val logger = getLogger(classOf[Http4sServlet[F]])