+++ servlet/src/main/scala/org/http4s/servlet/syntax/ServletContextSyntax.scala
@@ -6,7 +6,8 @@ import javax.servlet.{ServletContext, ServletRegistration}
 
 import cats.effect._
 import org.http4s.server.AsyncTimeoutSupport
-import org.http4s.util.threads.DefaultExecutionContext
+
+import scala.concurrent.ExecutionContext
 
 trait ServletContextSyntax {
   implicit def ToServletContextOps(self: ServletContext): ServletContextOps = new ServletContextOps(self)
@@ -14,11 +15,11 @@ trait ServletContextSyntax {
 
 final class ServletContextOps private[syntax](val self: ServletContext) extends AnyVal {
   /** Wraps an HttpService and mounts it as a servlet */
-  def mountService[F[_]: Effect](name: String, service: HttpService[F], mapping: String = "/*"): ServletRegistration.Dynamic = {
+  def mountService[F[_]: Effect](name: String, service: HttpService[F], mapping: String = "/*")(implicit ec: ExecutionContext = ExecutionContext.global): ServletRegistration.Dynamic = {
     val servlet = new Http4sServlet(
       service = service,
       asyncTimeout = AsyncTimeoutSupport.DefaultAsyncTimeout,
-      executionContext = DefaultExecutionContext,
+      executionContext = executionContext,
       servletIo = servletIo
     )
     val reg = self.addServlet(name, servlet)