+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# v0.17.0-M4 (unreleased)
+# v0.17.0-RC1 (expected week of 2017-07-31)
 * Port `ChunkAggregator` to fs2
 * Add logging middleware
 * Add `request#withUri`, which updates `request#pathInfo`. Remove
@@ -6,7 +6,7 @@
 * Standardize on `ExecutionContext` over `Strategy` and `ExecutorService`
 * Implement `Age` header
 * Fix `Client#toHttpService` to not dispose until the body is consumed
-* Change `Content-Length.apply` to `Contnet-Length.unsafeFromLong`
+* Change `Content-Length.apply` to `Content-Length.unsafeFromLong`
 * Add a buffered implementation of `EntityDecoder[Multipart]`
 * Remove `name` from `AttributeKey`. `AttributeKey`s are compared by reference,
   so the `name` was for display only.
@@ -36,12 +36,22 @@
 	* Client retry middleware can't check idempotence of requests
 	* Utilties in `org.http4s.util.io` not yet ported
 
-# v0.16.0-M4 (unreleased)
+# v0.16.0-RC1 (expected week of 2017-07-31)
 * Remove laziness from `ArbitraryInstances`
 * Support an arbitrary predicate for CORS allowed origins
 * Support `Access-Control-Expose-Headers` header for CORS
 * Fix thread safety issue in `EntityDecoder[XML]`
->>>>>>> release-0.16.x
+* Support IPV6 headers in `X-Forwarded-For`
+* Upgraded dependencies:
+    * async-http-client-2.0.33
+    * http4s-websocket-0.2.0
+    * jetty-9.4.6.v20170531
+    * json4s-3.5.3
+    * log4s-1.3.6
+    * metrics-core-3.2.3
+    * scalaz-7.2.14
+    * tomcat-8.5.19
+    * twirl-1.3.3
 
 # v0.16.0-M3 (2017-05-25)
 * Fix `WebjarService` so it matches assets.